Memtest86+: Set correct bounds when locating coreboot tables
When the coreboot tables are pointed to by a forward header, they are
often located at an address greater than 0x1000. However, when locating
these tables, the search end boundary was set to 0x1000. As a result,
the tables would not be found, and Memtest86+ would not start at all.
This happens, for example, on an ASUS P8H61-M LX.

To remedy this, change the end boundary of the search to be 0x1000
greater than the start.

Change-Id: I615672a8aa6253fb27c4f9c99101556b4a5a81a9
Signed-off-by: Tristan Corrick <tristan@corrick.kiwi>
Reviewed-on: https://review.coreboot.org/27730
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Martin Roth <martinroth@google.com>