Memtest86+: Handle missing SMBIOS information
Some information about the system is obtained from the SMBIOS table,
and is printed at the bottom of the screen. However, it was not checked
whether an entry was present before evaluating it, which led to a mess
of irrelevant characters printed to the screen.

This change is particularly relevant for coreboot, as the Socket
Designation field is not currently being filled in the SMBIOS table,
but Memtest86+ looks there for information.

Tested on an ASUS P8H61-M LX with a missing Socket Designation, as well
as the Socket Designation being present.

Change-Id: I6f063f5bf09279aafc9295cfcfccf98bc24a7ff0
Signed-off-by: Tristan Corrick <tristan@corrick.kiwi>
Reviewed-on: https://review.coreboot.org/27731
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Paul Menzel <paulepanter@users.sourceforge.net>
Reviewed-by: Martin Roth <martinroth@google.com>