Memtest86+: Support for configuring console from LB_SERIAL entry
I don't like how payload consoles are configured via Kconfig. This
leads to Kconfig needing to know things that would otherwise be the
responsibility of the code. We've had LB_SERIAL available for a very
long time, which does exactly this, but without Kconfig.

I complained to Martin, and he challenged me to fix this for Memtest86
Challenge accepted!

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>
Change-Id: Ifa920b8297c7edca53372482107a071b28296340
Reviewed-on: https://review.coreboot.org/c/memtest86plus/+/37056
Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-by: Stefan Reinauer <stefan.reinauer@coreboot.org>